requirements.txt
python-binance==1.0.16
pandas==2.0.0
numpy==1.24.3
scikit-learn==1.2.2
ta==0.10.2
pytest==7.3.1
python-dotenv==1.0.0
matplotlib==3.7.1
seaborn==0.12.2
jupyter==1.0.0
